refactor: Replace useNavigate when it's not needed
