+++ frontend/src/components/molecules/newChatButton.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
+import { redirect } from 'react-router-dom';
 
 import AddIcon from '@mui/icons-material/Add';
 import { Box } from '@mui/material';
@@ -9,7 +9,6 @@ import { AccentButton, useChat } from '@chainlit/components';
 import NewChatDialog from './newChatDialog';
 
 export default function NewChatButton() {
-  const navigate = useNavigate();
   const [open, setOpen] = useState(false);
   const { clear } = useChat();
 
@@ -23,7 +22,7 @@ export default function NewChatButton() {
 
   const handleConfirm = () => {
     clear();
-    navigate('/');
+    redirect('/');
     handleClose();
   };
